JWT Authentication.

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 // @desc Get goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user._id });
 
   res.status(200).json(goals);
 });
@@ -22,6 +23,7 @@ const setGoal = asyncHandler(async (req, res) => {
 
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user._id,
   });
 
   res.status(200).json(goal);
@@ -36,6 +38,20 @@ const updateGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(404);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user._id);
+
+  // Check for user
+  if (!user) {
+    res.status(404);
+    throw new Error("User not found");
+  }
+
+  // Check user is owner of goal
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized to update goal");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -53,6 +69,20 @@ const deleteGoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(404);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user._id);
+
+  // Check for user
+  if (!user) {
+    res.status(404);
+    throw new Error("User not found");
+  }
+
+  // Check user is owner of goal
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("User not authorized to delete goal");
   }
 
   const deleteGoal = await Goal.findByIdAndDelete(req.params.id);
